Fix animation system by making the timer a float instead of an unsiged int - rounding caused choppy animation

+++ TheDarkBelow/Systems/AnimationSystem.h
@@ -15,7 +15,7 @@ namespace DarkBelow {
 			void update(float dt);
 
 		private:
-			size_t timer;
+			float timer;
 		};
 	}
 }
